LayoutTests/platform/ios-simulator/accessibility/text-marker-validation-expected.txt

element 1
element 2
element 3
element 4
This test ensures that a text range can be found for an element within a web page.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS succeeded is true
PASS successfullyParsed is true

TEST COMPLETE
